==> notebooks/02-goal_option_1.ipynb <==
# Goals Option 1

## How much are nearby Hospitals and ASCs paid for the same procedure?

Filters for Physician/Other Supplier
- Entity Type = O
- Provider Type = Ambulatory Surgical Center
- CPT = 43249

Filters for Hospital Outpatient
- APC = 5302

Compare results within a CBSA

# Stretch Goals

## Extend to other HCPCS/APCs
- Best place to start is to find an APC that is well-populated in the Hospital file
- Next find a well-populated HCPCS from the APC in the Physician/Other Supplier File

## Compare Doctor Utilization for a procedure
- Filters for Physician/Other Supplier
  - Entity Type = I
  - Place of Service = O
- HCPCS = 99213 (or any office visit – 99211-99215 / 99201-99205)
- How many different patients does each specialty typically see in a year? How does that vary by CBSA?

# To Do List

- Create full_name column for providers and outpatients(?)
- classification1 dataframe setup is broken, can't import it

# import libraries
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

# set up dataframes
providers_tsv = '../data/Medicare_Provider_Util_Payment_PUF_CY2017.tsv'
outpatient_excel = '../data/MUP_OHP_R19_P04_V10_D17_APC_Provider.xlsx'
classification1_csv = '../data/508-Compliant-Version-of-2020_january_web_addendum_b.12312019.csv'
classification2_excel = '../data/2020_january_web_addendum_b.12312019.xlsx'
cbsa_excel = '../data/ZIP_CBSA_032020.xlsx'

# NOTE: providers_csv is a TAB-DELIMITED file, Use sep='\t'
providers = pd.read_csv(providers_tsv, sep='\t', low_memory=False) 
outpatient = pd.read_excel(outpatient_excel, skiprows=5) 
# classification1 = pd.read_csv(classification1_csv, engine='python') 
# classification2 = pd.read_excel(classification2_excel) 
classifications = pd.read_excel(classification2_excel) 
cbsa = pd.read_excel(cbsa_excel) 

# NOTE: the first line in providers_tsv is copyright info
outpatient.head()

# Inner join on HCPS code will eliminate any duplicates
# NOTE: Trailing spaces were found in column names, they're included below and will be removed later
# memory usage after join only increased by 0.5 MB

# classifications = pd.merge(left=classification1, right=classification2, 
#                            how='inner', 
#                            on=['HCPCS Code','Short Descriptor','SI','APC ',
#                                'Relative Weight','Payment Rate ','National Unadjusted Copayment ',
#                               'Minimum Unadjusted Copayment '])
# classifications.info()

# Format Dataframes
- clean up column names and make lowercase
- make provider names and cities lowercase as well

# column names contain newline characters, remove them
# also convert spaces to underscores and make lowercase
outpatient.columns = outpatient.columns.str.replace('\n', '_').str.replace(' ', '_').str.lower()
providers.columns = providers.columns.str.replace(' ', '_').str.lower()
classifications.columns = classifications.columns.str.replace(' ', '_').str.lower()
cbsa.columns = cbsa.columns.str.replace(' ', '_').str.lower()

providers.head()

# make provider names and cities lowercase as well
providers['nppes_provider_last_org_name'] = providers['nppes_provider_last_org_name'].str.capitalize()
providers['nppes_provider_first_name'] = providers['nppes_provider_first_name'].str.capitalize()
providers['nppes_provider_city'] = providers['nppes_provider_city'].str.capitalize()


providers.head()



# Join Zip to CBSA table with both providers AND outpatient
- Note: when I convert datatype here, it breaks due to Canadian postal codes. Move this section to after non-Tenn. data has been filtered out

# save unfiltered files for later use
# outpatient.to_csv('../data/outpatient_clean.csv')
# providers.to_csv('../data/providers_clean.csv')


# providers['nppes_provider_zip'] = pd.to_numeric(providers['nppes_provider_zip'])

# providers.info()


outpatient.head()

# Create Full_Name Column in Providers

## This needs to be added before the individual dataframes are set up
- IF Entity_Type = O THEN Last_Name
- IF Entity_Type = I THEN First_Name + “ “ + Middle Initial + “ “ + Last_Name + “, “ + Credentials


# create full_name column for providers
# df['Price'] = [1500 if x =='Music' else 800 for x in df['Event']] 


# if x > 10:
#   print("Above ten,")
#   if x > 20:
#     print("and also above 20!")
#   else:
#     print("but not above 20.")

    
# if providers['nppes_entity_code'] == 'O':
#     providers['full_name'] = providers['nppes_provider_last_org_name']
#     if providers['nppes_entity_code'] == 'I':
#         providers['full_name'] = providers['nppes_provider_first_name'] + ' '
#         + providers['nppes_provider_mi'] + ' ' + providers['nppes_provider_last_org_name']
#     else:
#         providers['full_name'] = NULL
        


# providers['full_name'] = [nppes_provider_last_org_name 
#                            if x == 'O'
# #                            else nppes_provider_first_name + ' ' 
# #                               + nppes_provider_mi + ' ' 
# #                               + nppes_provider_last_org_name
# #                            for x in providers['nppes_entity_code']
#                            else providers['nppes_provider_first_name'] + ' ' 
#                               + providers['nppes_provider_mi'] + ' ' 
#                               + providers['nppes_provider_last_org_name']
#                            for x in providers['nppes_entity_code']
#                           ]
# providers.head()

# Set up City Dataframes

We're examining 3 cities: Nashville, Los Angeles, New York City. Starting with Nashville, we set up dataframes and take a look at the procedures done there.

# set up Nashville Outpatient Dataframe
tennessee_outpatient = outpatient[outpatient ['provider_state'] == 'TN'] 
nashville_outpatient = tennessee_outpatient[tennessee_outpatient ['provider_city'] == 'Nashville'] 
nashville_outpatient

# set up Nashville Providers Dataframe
tennessee_providers = providers[providers ['nppes_provider_state'] == 'TN'] 
nashville_providers = tennessee_providers[tennessee_providers ['nppes_provider_city'] == 'Nashville'] 
nashville_providers

# Set up New York City and Los Angeles dataframes
ny_outpatient = outpatient[outpatient ['provider_state'] == 'NY'] 
nyc_outpatient = ny_outpatient[ny_outpatient ['provider_city'] == 'New York'] 
nyc_outpatient

ca_outpatient = outpatient[outpatient ['provider_state'] == 'CA'] 
la_outpatient = ca_outpatient[ca_outpatient ['provider_city'] == 'Los Angeles'] 
la_outpatient

ny_providers = providers[providers ['nppes_provider_state'] == 'NY'] 
nyc_providers = ny_providers[ny_providers ['nppes_provider_city'] == 'New York'] 
nyc_providers

ca_providers = providers[providers ['nppes_provider_state'] == 'CA'] 
la_providers = ca_providers[ca_providers ['nppes_provider_city'] == 'Los Angeles'] 
la_providers

# Filter dataframes

## Filter outpatient
# APC = 5302

# individual statements for filters
tennessee_outpatient = tennessee_outpatient.loc[
    tennessee_outpatient['apc'] == 5302
]

nashville_outpatient = nashville_outpatient.loc[
    nashville_outpatient['apc'] == 5302
]

nyc_outpatient = nyc_outpatient.loc[
    nyc_outpatient['apc'] == 5302
]

la_outpatient = la_outpatient.loc[
    la_outpatient['apc'] == 5302
]

# nyc_outpatient.head()

## Filter providers
# SELECT entity type = 'O'
# Provider Type = Ambulatory Surgical Center
# CPT/HCPCS = 43249
# DEBUGDEBUG: I'm getting zero rows when I add the hcpcs filter <-------------------------------------


# individual statements for filters
tennessee_providers = tennessee_providers.loc[
    (tennessee_providers['nppes_entity_code'] == 'O') &
    (tennessee_providers['provider_type'] == 'Ambulatory Surgical Center')
   & (tennessee_providers['hcpcs_code'] == '43249')
]

nashville_providers = nashville_providers.loc[
    (nashville_providers['nppes_entity_code'] == 'O') &
    (nashville_providers['provider_type'] == 'Ambulatory Surgical Center')
   & (nashville_providers['hcpcs_code'] == '43249')
]

nyc_providers = nyc_providers.loc[
    (nyc_providers['nppes_entity_code'] == 'O') &
    (nyc_providers['provider_type'] == 'Ambulatory Surgical Center')
    & (nyc_providers['hcpcs_code'] == '43249')
]

la_providers = la_providers.loc[
    (la_providers['nppes_entity_code'] == 'O') &
    (la_providers['provider_type'] == 'Ambulatory Surgical Center')
   & (la_providers['hcpcs_code'] == '43249')
]

tennessee_providers.head()

# Add CBSA info

# convert zip code to numeric fields
tennessee_providers['nppes_provider_zip'] = pd.to_numeric(tennessee_providers['nppes_provider_zip'])
tennessee_outpatient['provider_zip_code'] = pd.to_numeric(tennessee_outpatient['provider_zip_code'])

tennessee_providers = pd.merge(left=tennessee_providers, right=cbsa, left_on='nppes_provider_zip', right_on='zip', how='outer')
tennessee_outpatient = pd.merge(left=tennessee_outpatient, right=cbsa, left_on='provider_zip_code', right_on='zip', how='outer')

tennessee_outpatient.info()

tennessee_providers.info()


# These Dataframes are Loaded and Ready for Use

## Outpatients
- tennessee_outpatient
- nashville_outpatient
- nyc_outpatient
- la_outpatient

## Providers
- tennessee_providers
- nashville_providers
- nyc_providers
- la_providers

# Next Steps
If no more processing, time to save these in new files for analysis
- tennessee_outpatient_clean
- nashville_outpatient_clean
- nyc_outpatient_clean
- la_outpatient_clean
- nashville_providers_clean
- nyc_providers_clean
- la_providers_clean


# save clean dataframes
tennessee_outpatient.to_csv('../data/tennessee_outpatient_clean.csv')
nashville_outpatient.to_csv('../data/nashville_outpatient_clean.csv')
tennessee_providers.to_csv('../data/tennessee_providers_clean.csv')
nashville_providers.to_csv('../data/nashville_providers_clean.csv')

# cbsa.to_csv('../data/cbsa_clean.csv')
# nyc_outpatient.to_csv('../data/nyc_outpatient_clean.csv')
# la_outpatient.to_csv('../data/la_outpatient_clean.csv')
# nyc_providers.to_csv('../data/nyc_providers_clean.csv')
# la_providers.to_csv('../data/la_providers_clean.csv')


nashville_outpatient

nashville_providers.info()
