==> notebooks/03-goal_option_1.ipynb <==
# import libraries
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

# import and set up dataframes

# tennessee_outpatient.to_csv('../data/tennessee_outpatient_clean.csv')
# nashville_outpatient.to_csv('../data/nashville_outpatient_clean.csv')
# tennessee_providers.to_csv('../data/tennessee_providers_clean.csv')
# nashville_providers.to_csv('../data/nashville_providers_clean.csv')

cbsa = pd.read_excel('../data/ZIP_CBSA_032020.xlsx') 

#   Outpatients
tennessee_outpatient_clean = pd.read_csv('../data/tennessee_outpatient_clean.csv', low_memory=False)
# nashville_outpatient_clean = pd.read_csv('../data/nashville_outpatient_clean.csv')
# nyc_outpatient_clean = pd.read_csv('../data/nyc_outpatient_clean.csv')
# la_outpatient_clean = pd.read_csv('../data/la_outpatient_clean.csv')

#   Providers
tennessee_providers_clean = pd.read_csv('../data/tennessee_providers_clean.csv', )
# nashville_providers_clean = pd.read_csv('../data/nashville_providers_clean.csv')
# nyc_providers_clean = pd.read_csv('../data/nyc_providers_clean.csv')
# la_providers_clean = pd.read_csv('../data/la_providers_clean.csv')


tennessee_outpatient_clean.info()

tennessee_providers_clean.info()

# Note
This looks like a small data set for Nashville. Concerned that it was too small, I went back to 02-goal_option_1 and confirmed that this small set is correct. It is the set of outpatients filtered by APC = 5302.

# nashville_outpatient_clean

# Convert zipcodes from float to int


# drop NaN
tennessee_providers_clean['nppes_provider_zip']

tennessee_providers_clean = tennessee_providers_clean.dropna()


# # tennessee_providers_clean['nppes_provider_zip'] = tennessee_providers_clean['nppes_provider_zip'].dropna()
# # tennessee_outpatient_clean['provider_zip_code'] = tennessee_outpatient_clean['provider_zip_code'].dropna()

# tennessee_providers_clean = pd.merge(left=tennessee_providers_clean, right=cbsa, left_on='nppes_provider_zip', right_on='ZIP', how='outer')
# tennessee_outpatient_clean = pd.merge(left=tennessee_outpatient_clean, right=cbsa, left_on='provider_zip_code', right_on='ZIP', how='outer')

# tennessee_outpatient_clean.info()

# convert zipcodes from float to int
# .astype(int)
# tennessee_providers_clean['nppes_provider_zip'] = tennessee_providers_clean['nppes_provider_zip'].astype(int)
# tennessee_outpatient_clean['provider_zip_code'] = tennessee_outpatient_clean['provider_zip_code'].astype(int)

tennessee_providers_clean['nppes_provider_zip']
# tennessee_outpatient_clean['provider_zip_code']

# tennessee_outpatient_clean.info()

# Get average_medicare_allowed_amount

## NOTE: the Nashville dataframes are redundant as we're looking at Tenn by CBSA. They're included below, but we'll only use the Tennessee dataframes as Nashville is included

# outpatient: average_medicare_allowed_amount
# nashville_outpatient_amaa = nashville_outpatient_clean[['provider_zip_code', 'provider_name', 'average_medicare_allowed_amount']].sort_values('average_medicare_allowed_amount')
# nashville_outpatient_amaa

# provider: average_medicare_allowed_amount
# nashville_providers_amaa = nashville_providers_clean[['cbsa', 'nppes_provider_zip', 'nppes_provider_last_org_name', 'average_medicare_allowed_amt']].sort_values('average_medicare_allowed_amt')
# nashville_providers_amaa

# outpatient: average_medicare_allowed_amount
tennessee_outpatient_amaa = tennessee_outpatient_clean[['cbsa', 'provider_zip_code', 'provider_name', 'average_medicare_allowed_amount']].sort_values('average_medicare_allowed_amount')
tennessee_outpatient_amaa

tennessee_providers_amaa = tennessee_providers_clean[['cbsa','nppes_provider_zip', 'nppes_provider_last_org_name', 'average_medicare_allowed_amt']].sort_values('average_medicare_allowed_amt')
tennessee_providers_amaa

# Save Dataframe

s for further analysis

# # save clean dataframes
tennessee_outpatient_amaa.to_csv('../data/tennessee_outpatient_amaa.csv')
tennessee_providers_amaa.to_csv('../data/tennessee_providers_amaa.csv')
# nashville_outpatient_amaa.to_csv('../data/nashville_outpatient_amaa.csv')
# nashville_providers_amaa.to_csv('../data/nashville_providers_amaa.csv')


tennessee_outpatient_clean.to_csv('../data/tennessee_outpatient_filtered.csv')
tennessee_providers_clean.to_csv('../data/tennessee_providers_filtered.csv')
