==> notebooks/04-goal_option_1.ipynb <==
# Analyzing average_medicare_allowed_amount in Tennessee Dataframes

# import libraries
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

# import and set up dataframes
# tennessee_outpatient_amaa = pd.read_csv('../data/tennessee_outpatient_amaa.csv')
# tennessee_providers_amaa = pd.read_csv('../data/tennessee_providers_amaa.csv')
tennessee_outpatient_clean = pd.read_csv('../data/tennessee_outpatient_clean.csv', low_memory=False)
tennessee_providers_clean = pd.read_csv('../data/tennessee_providers_clean.csv', low_memory=False)
cbsa_to_name = pd.read_csv('../data/core-based-statistical-areas-cbsas-and-combined-statistical-areas-csas.csv', engine='python')

# tennessee_outpatient_filtered = pd.read_csv('../data/tennessee_outpatient_filtered.csv')
# tennessee_providers_filtered = pd.read_csv('../data/tennessee_providers_filtered.csv')

tennessee_outpatient_clean.head()

tennessee_outpatient_clean.info()

tennessee_providers_clean.head(25)

hello world


tennessee_providers_clean.info()

# join CBSA2name so that we get CBSA names

cbsa_to_name.info()

cbsa_to_name.columns = cbsa_to_name.columns.str.replace(' ', '_').str.lower()
cbsa_to_name['cbsa_code'] = pd.to_numeric(cbsa_to_name['cbsa_code'], errors='coerce')
cbsa_to_name.info()

tennessee_outpatient_clean = pd.merge(left=tennessee_outpatient_clean, right=cbsa_to_name, 
                                      left_on='cbsa', right_on='cbsa_code', how='outer')
tennessee_outpatient_clean.info()

# Get Files Ready for Tableau
- drop un-needed columns
- fix datatypes

# Drop all but listed tables for outpatient
tennessee_outpatient_tableau = tennessee_outpatient_clean[[
    'provider_name',
    'apc',
    'apc_description',
    'beneficiaries',
    'comprehensive_apc_services',
    'average_estimated_total_submitted_charges',
    'average_medicare_allowed_amount',
    'average_medicare_payment_amount',
    'average_medicare_outlier_amount',
    'zip',
    'cbsa',
    'cbsa_title']]
tennessee_outpatient_tableau = tennessee_outpatient_tableau.dropna()
tennessee_outpatient_tableau

tennessee_outpatient_tableau['zip']= tennessee_outpatient_tableau['zip'].astype(int)
tennessee_outpatient_tableau['apc']= tennessee_outpatient_tableau['apc'].astype(int)
tennessee_outpatient_tableau['cbsa']= tennessee_outpatient_tableau['cbsa'].astype(int)
tennessee_outpatient_tableau.info()

tennessee_outpatient_tableau

# Drop all but listed tables for proviers
tennessee_providers_tableau = tennessee_providers_clean[[
    'nppes_provider_last_org_name',
    'provider_type',
    'place_of_service',
    'hcpcs_code',
    'hcpcs_description',
    'line_srvc_cnt',
    'bene_unique_cnt',
    'average_medicare_allowed_amt',
    'average_submitted_chrg_amt',
    'average_medicare_payment_amt',
    'average_medicare_standard_amt']]
tennessee_providers_tableau = tennessee_providers_tableau.dropna()
tennessee_providers_tableau.head(50)

tennessee_providers_tableau['hcpcs_code']= tennessee_providers_tableau['hcpcs_code'].astype(int)
tennessee_providers_tableau['hcpcs_code']= tennessee_providers_tableau['hcpcs_code'].astype(object)
tennessee_providers_tableau.info()

# save files

tennessee_outpatient_tableau.to_csv('../data/tennessee_outpatient_tableau.csv')
tennessee_providers_tableau.to_csv('../data/tennessee_providers_tableau.csv')


tennessee_outpatient_tableau
